requirements.txt
Django>=5.1.0,<5.2.0
django-money>=3.4.0,<3.5.0
djangorestframework>=3.15.0,<3.16.0
fakeredis>=2.27.0,<3.0.0
fastapi>=0.115.0,<0.116.0
hypothesis>=6.108.0,<7.0.0
jsonschema>=4.23.0,<5.0.0
networkx>=3.4.0,<4.0.0
openapi-spec-validator>=0.7.0,<0.8.0
orjson>=3.10.0,<4.0.0
pusher>=3.3.0,<4.0.0
pydantic>=2.10.0,<3.0.0
rich>=13.9.0,<14.0.0
psycopg2>=2.9.0,<3.0.0
django-redis>=5.4.0,<6.0.0
django-cors-headers>=4.7.0,<5.0.0
django-zen-queries>=2.1.0,<3.0.0
cytoolz>=1.0.0,<2.0.0